uploaded entire clipping site
tried to optimize for smaller screen, on mobile is okay-ish. Screen reduction needs to be dealt with. Biggest problem is elements that just won't shrink in a smart way.
Also, adding properties that are useless on full page but serve a purpose on shrinked screens is really a big deal